hyperv: don't apply checkpoints on the first run and retry to remove snaspshots

+++ Libraries/HyperV.psm1
@@ -185,11 +185,17 @@ Function Delete-HyperVGroup([string]$HyperVGroupName, [string]$HyperVHost, $Setu
             $cleanupDone--
             return
         }
-        Remove-VMSnapshot -VMName $vm.Name -ComputerName $HyperVHost `
-            -IncludeAllChildCheckpoints -Confirm:$false -ErrorAction SilentlyContinue
-        if (!$?) {
-            Write-LogErr ("Failed to remove snapshots for VM {0}" -f @($vm.Name))
-            return $false
+        $retriesCleanSnapshots = 0
+        $maxRetriesCleanSnapshots = 3
+        while ($retriesCleanSnapshots -lt $maxRetriesCleanSnapshots) {
+            Remove-VMSnapshot -VMName $vm.Name -ComputerName $HyperVHost `
+                -IncludeAllChildCheckpoints -Confirm:$false -ErrorAction SilentlyContinue
+            if ($?) {
+                Write-LogWarn ("Failed to remove snapshots for VM {0}. Retrying..." -f @($vm.Name))
+                $retriesCleanSnapshots++
+            } else {
+                break
+            }
         }
         Wait-VMStatus -VMName $vm.Name -VMStatus "Operating Normally" -RetryInterval 2 `
             -HvServer $HyperVHost
@@ -207,13 +213,15 @@ Function Delete-HyperVGroup([string]$HyperVGroupName, [string]$HyperVHost, $Setu
             }
             Invoke-Command @invokeCommandParams
             if (!$?) {
-                Write-LogInfo "Failed to remove ${vhdPath} using Invoke-Command"
                 $vhdUncPath = $vhdPath -replace '^(.):', "\\${HyperVHost}\`$1$"
-                Write-LogInfo "Removing ${vhdUncPath} ..."
-                Remove-Item -Path $vhdUncPath -Force
-                if (!$? -or (Test-Path $vhdUncPath)) {
-                    Write-LogErr "Failed to remove ${vhdPath} using UNC paths"
-                    return $false
+                if ((Test-Path $vhdUncPath)) {
+                    Write-LogWarn "Failed to remove ${vhdPath} using Invoke-Command"
+                    Write-LogInfo "Removing ${vhdUncPath} ..."
+                    Remove-Item -Path $vhdUncPath -Force
+                    if (!$? -or (Test-Path $vhdUncPath)) {
+                        Write-LogErr "Failed to remove ${vhdPath} using UNC paths"
+                        return $false
+                    }
                 }
             }
             Write-LogInfo "VHD ${vhdPath} removed!"
